+++ lib/localeapp/missing_translations.rb
@@ -2,11 +2,20 @@ module Localeapp
   MissingTranslationRecord = Struct.new(:key, :locale, :description, :options)
 
   class MissingTranslations
+    @cached_keys = []
+
+    class << self
+      attr_accessor :cached_keys
+    end
+
     def initialize
       @translations = Hash.new { |h, k| h[k] = {} }
     end
 
     def add(locale, key, description = nil, options = {})
+      separator = options.delete(:separator) { I18n.default_separator }
+      scope = options.delete(:scope)
+      key = I18n.normalize_keys(nil, key, scope, separator).map(&:to_s).join(separator)
       record = MissingTranslationRecord.new(key, locale, description, options)
       @translations[locale][key] = record
     end
@@ -15,23 +24,33 @@ module Localeapp
       @translations[locale]
     end
 
-    # This method will get cleverer so we don't resend keys we've
-    # already sent, or send multiple times for the same locale etc.
-    # For now it's pretty dumb
     def to_send
       data = []
       # need the sort to make specs work under 1.8
       @translations.sort { |a, b| a.to_s <=> b.to_s }.each do |locale, records|
         records.each do |key, record|
-          missing_data = {}
-          missing_data[:key] = key
-          missing_data[:locale] = locale
+          next if cached?(key)
+          cache(key)
+          missing_data = {:key => key, :locale => locale, :options => record.options}
           missing_data[:description] = record.description if record.description
-          missing_data[:options] = record.options
           data << missing_data
         end
       end
       data
+    end
+
+    private
+
+    def cached_keys
+      self.class.cached_keys
+    end
+
+    def cached?(key)
+      Localeapp.configuration.cache_missing_translations && cached_keys.include?(key)
+    end
+
+    def cache(key)
+      cached_keys << key if Localeapp.configuration.cache_missing_translations
     end
   end
 end