+++ lib/localeapp/rails.rb
@@ -1,3 +1,5 @@
+require 'fileutils'
+
 module Localeapp
   module Rails
     def self.initialize
@@ -36,8 +38,10 @@ module Localeapp
     end
 
     def self.initialize_synchronization_data_file
-      if !File.exists?(Localeapp.configuration.synchronization_data_file)
-        File.open(Localeapp.configuration.synchronization_data_file, 'w') do |f|
+      sync_file = Localeapp.configuration.synchronization_data_file
+      if !File.exists?(sync_file)
+        FileUtils.mkdir_p(File.dirname(sync_file))
+        File.open(sync_file, 'w') do |f|
           f.write({:polled_at => Time.now.to_i, :updated_at => Time.now.to_i}.to_yaml)
         end
       end