+++ spec/localeapp/exception_handler_spec.rb
@@ -20,17 +20,28 @@ describe Localeapp::ExceptionHandler, '#call(exception, locale, key, options)' d
   end
 
   it "delegates to super for handling of output" do
-    expect(I18n.t('foo', rescue_format: :html)).to eq("translation missing: foo")
+    expect(I18n.t('foo')).to eq("translation missing: en.foo")
   end
 
   it "delegates to super for handling of rescue_format" do
-    expect(I18n.t('foo', rescue_format: :html)).to eq("<span class=\"translation_missing\" title=\"translation missing: en.foo\">Foo</span>")
+    expect(I18n.t('foo', :rescue_format => :html)).to eq("<span class=\"translation_missing\" title=\"translation missing: en.foo\">Foo</span>")
+  end
+
+  it "handles when the default is a Symbol that can be resolved" do
+    I18n.backend.store_translations(:en, {:default_symbol_test => 'is resolved'})
+    Localeapp.missing_translations.should_not_receive(:add)
+    I18n.t(:foo, :default => :default_symbol_test).should == 'is resolved'
+  end
+
+  it "handles when the default is a Symbol that can't be resolved" do
+    Localeapp.missing_translations.should_receive(:add).with(:en, :foo, nil, {:default => :bar})
+    I18n.t(:foo, :default => :bar)
   end
 
   it "handles missing translation exception" do
     expect {
       exception = Localeapp::I18nMissingTranslationException.new(:en, 'foo', {})
-      Localeapp::ExceptionHandler.new.call(exception, :en, 'foo', {})
+      I18n::ExceptionHandler.new.call(exception, :en, 'foo', {})
     }.to_not raise_error
   end
 end