+++ spec/localeapp/missing_translations_spec.rb
@@ -9,6 +9,14 @@ describe Localeapp::MissingTranslations, "#add(locale, key, description = nil, o
     translations[:en]['foo'].description.should == 'bar'
     translations[:en]['foo'].options.should == { :baz => 'bam' }
   end
+
+  it "respects I18n options when constructing a key" do
+    translations = Localeapp::MissingTranslations.new
+    translations.add(:en, 'foo', nil, { :scope => 'bar', :baz => 'bam' })
+
+    translations[:en].should include('bar.foo')
+    translations[:en]['bar.foo'].options.should == { :baz => 'bam' }
+  end
 end
 
 describe Localeapp::MissingTranslations, "#to_send" do
@@ -27,5 +35,29 @@ describe Localeapp::MissingTranslations, "#to_send" do
     to_send[1][:locale].should == :es
     to_send[1][:description].should == 'baz'
     to_send[1][:options].should == {}
+  end
+
+  it "doesn't send the same key twice when cache_missing_translations is true" do
+    with_configuration(:cache_missing_translations => true) do
+      translation_a = Localeapp::MissingTranslations.new
+      translation_a.add(:es, 'foobybar')
+      translation_a.to_send.size.should == 1
+
+      translation_b = Localeapp::MissingTranslations.new
+      translation_b.add(:en, 'foobybar')
+      translation_a.to_send.size.should == 0
+    end
+  end
+
+  it "can send the same key twice when cache_missing_translations is false" do
+    with_configuration(:cache_missing_translations => false) do
+      translation_a = Localeapp::MissingTranslations.new
+      translation_a.add(:es, 'foobybar')
+      translation_a.to_send.size.should == 1
+
+      translation_b = Localeapp::MissingTranslations.new
+      translation_b.add(:en, 'foobybar')
+      translation_a.to_send.size.should == 1
+    end
   end
 end